fix: finish checkout endpoint in cart & management and view orders

+++ shoppingCart/views.py
@@ -1,16 +1,27 @@
 from django.db import IntegrityError
-from django.shortcuts import render
+from django.shortcuts import get_object_or_404, render
 from rest_framework.decorators import api_view, permission_classes
 from rest_framework.response import Response
 from rest_framework.permissions import IsAuthenticated
 from rest_framework import status
 from products.models import Product
+from user.models import User
 from .models import Cart , CartItem 
-from orderTest.models import Order2
-from orderTest.serializer import Order2Serializer
 from .serializers import CartItemSerializer , CartSerializer
+from orders.models import Order , OrderItem
+from orders.serializers import OrderItemsSerializer , OrderSerializer
 
 # Create your views here.
+def get_cart_item(user):
+    try:
+        cart = Cart.objects.get(user=user)
+        cart_items = cart.cartitem_set.filter(status='onCart') 
+        serializer = CartItemSerializer(cart_items, many=True)
+        return serializer.data
+    except Cart.DoesNotExist:
+        return []
+    except Exception as e:
+        return []
 
 @api_view(['GET'])
 @permission_classes([IsAuthenticated])
@@ -50,8 +61,8 @@ def add_item(request):
         else:
             cart_item.quantity = quantity
             
-        product.quantity = product.quantity - quantity
-        product.save()
+        # product.quantity = product.quantity - quantity
+        # product.save()
         cart_item.save()
             
         serializer = CartItemSerializer(cart_item)
@@ -68,8 +79,8 @@ def remove_item(request, cart_item_id):
     except CartItem.DoesNotExist:
         return Response({"error": "Cart item does not exist"}, status=status.HTTP_404_NOT_FOUND)
     
-    cart_item.product.quantity += cart_item.quantity
-    cart_item.product.save()
+    # cart_item.product.quantity += cart_item.quantity
+    # cart_item.product.save()
     cart_item.delete()
     return Response({" Deleted Successfully "},status=status.HTTP_204_NO_CONTENT)
 
@@ -132,23 +143,42 @@ def change_cart_item_status(request, cart_item_id):
     except Exception as e:
         return Response({'detail': str(e)}, status=status.HTTP_500_INTERNAL_SERVER_ERROR)
 
-
-@api_view(['POST'])
+@api_view(['GET'])
 @permission_classes([IsAuthenticated])
-
 def checkout(request):
     try:
-        user_cart = Cart.objects.get(user=request.user)
+        user = request.user
+        user_cart = get_object_or_404(Cart, user=user)
     except Cart.DoesNotExist:
         return Response({'detail': 'Cart not found'}, status=status.HTTP_404_NOT_FOUND)
-        
-    order = Order2.objects.create(user=request.user, status='pending')
-        
-    for cart_item in user_cart.cartitem_set.all():
-        order.cart_items.add(cart_item)
-        cart_item.status= "done"
-        
-    user_cart.cartitem_set.all().delete()  
+
+    items = get_cart_item(user)
+    user_instance = User.objects.get(email=user.email)
     
-    serializer = Order2Serializer(order)
-    return Response(serializer.data, status=status.HTTP_201_CREATED)
+    order_serializer = OrderSerializer(data={'user': user_instance.pk, 'status': 'pending'})
+
+    if order_serializer.is_valid():
+        order = order_serializer.save()  
+
+        for item in items:
+            order_item_data = {
+                'order': order.pk,  
+                'product': item['product']['product_id'],  
+                'quantity': item['quantity']  
+            }
+            order_item_serializer = OrderItemsSerializer(data=order_item_data)
+            if order_item_serializer.is_valid():
+                order_item_serializer.save()  
+                product_order = Product.objects.get(product_id = item['product']['product_id'])
+                print(product_order)
+                product_order.quantity = product_order.quantity - item['quantity']
+                product_order.save()
+                item['status'] = 'done'
+            else:
+                return Response(order_item_serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+
+        user_cart.cartitem_set.all().delete()
+
+        return Response(order_serializer.data, status=status.HTTP_201_CREATED)
+    else:
+        return Response(order_serializer.errors, status=status.HTTP_400_BAD_REQUEST)
